trying to make one of the sequential models compatible with akida
The steps to making this sequential sub-model compatible with Akida will be different for other sequential sub-models since they have different quirks. I think I tested it with the first big branch, then the left and right branches (check image for diagram of the model), but not the last part after concatenation.